+++ arch/xtensa/core/coredump.c
@@ -6,7 +6,7 @@
 
 #include <string.h>
 #include <zephyr/debug/coredump.h>
-#include <xtensa-asm2-context.h>
+#include <xtensa_asm2_context.h>
 #include <offsets.h>
 
 #define ARCH_HDR_VER			1
