+++ arch/xtensa/core/startup/reset_vector.S
