+++ arch/xtensa/core/xtensa_asm2_util.S
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include <xtensa-asm2-s.h>
+#include <xtensa_asm2_s.h>
 #include <offsets.h>
 #include <zsr.h>
